// src/types/types.tsx
// Generated by https://quicktype.io

//Generic Results
export interface GenericResultsType {
  count: number;
  next: string | null;
  previous: string | null;
  results: Item[];
}

// Items
export interface Item {
  name: string;
  title?: string;
  gender?: Gender;
  height: number;
  url: string;
}

export enum Gender {
  Female = 'female',
  Male = 'male',
  NA = 'n/a',
}